+++ app/Http/Controllers/NewsController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\News;
 use Validator;
-
+use DB;
 
 class NewsController extends Controller
 {
@@ -16,6 +16,9 @@ class NewsController extends Controller
      */
     public function index($bank_id)
     {
+        if(DB::table('bank')->find(['id' => $bank_id])==null){
+            return response()->json(["message"=>'Bank with id '.$bank_id. ' not found'],400); 
+        }
         $news = News::where('bank_id', $bank_id)->get();
         return response()->json($news,200);
     }
@@ -38,30 +41,37 @@ class NewsController extends Controller
      */
     public function store(Request $request,$bank_id)
     {
-        $rules = [
-            'content' => 'required|min:20',
-            'title' => 'required|min:4',
-            'specialist_id' => 'required|min:1',
-            'expires_in' => 'required|date',
-        ];
-        
-        $validator = Validator::make($request->all(), $rules);
-
-
-
-        if($validator->fails()){
-            return response()->json(["message"=>$validator->errors()],400); 
+        if($this->checkIfSuperrior(2)){
+            $rules = [
+                'content' => 'required|min:20',
+                'title' => 'required|min:4',
+                'specialist_id' => 'required|min:1',
+                'expires_in' => 'required|date',
+            ];
+            
+            $validator = Validator::make($request->all(), $rules);
+    
+            
+            if(DB::table('bank')->find(['id' => $bank_id])==null){
+                return response()->json(["message"=>'Bank with id '.$bank_id. ' not found'],400); 
+            }
+            if($validator->fails()){
+                return response()->json(["message"=>$validator->errors()],400); 
+            }
+    
+            $news = News::create([
+                'content' => request('content'),
+                'title' => request('title'),
+                'specialist_id' => request('specialist_id'),
+                'expires_in' => request('expires_in'),
+                'bank_id' => $bank_id
+            ]);
+          
+            return response()->json($news,201);
+        }else{
+            return response()->json(["message"=>'Your user do not have permissions to do this'],400); 
         }
-
-        $news = News::create([
-    		'content' => request('content'),
-    		'title' => request('title'),
-            'specialist_id' => request('specialist_id'),
-            'expires_in' => request('expires_in'),
-            'bank_id' => $bank_id
-    	]);
-      
-        return response()->json($news,201);
+       
     }
 
     /**
@@ -99,13 +109,17 @@ class NewsController extends Controller
      */
     public function update(Request $request,$bank_id, $id)
     {
-        $news = News::where('bank_id', $bank_id)->find($id);
-        if(is_null($news)){
-            return response()->json(["message"=>"Record not found"],404); 
+        if($this->checkIfSuperrior(2)){
+            $news = News::where('bank_id', $bank_id)->find($id);
+            if(is_null($news)){
+                return response()->json(["message"=>"Record not found"],404); 
+            }
+          
+            $news->update($request->all());
+            return response()->json($news, 200);
+        }else{
+            return response()->json(["message"=>'Your user do not have permissions to do this'],400); 
         }
-      
-        $news->update($request->all());
-        return response()->json($news, 200);
     }
 
     /**
@@ -116,11 +130,23 @@ class NewsController extends Controller
      */
     public function destroy($bank_id,$id)
     {
-        $news = News::where('bank_id', $bank_id)->find($id);
-        if(is_null($news)){
-            return response()->json(["message"=>"Record not found"],404); 
+        if($this->checkIfSuperrior(2)){
+            $news = News::where('bank_id', $bank_id)->find($id);
+            if(is_null($news)){
+                return response()->json(["message"=>"Record not found"],404); 
+            }
+            $news->delete();
+            return response()->json("DELETED", 204);
+        }else{
+            return response()->json(["message"=>'Your user do not have permissions to do this'],400); 
         }
-        $news->delete();
-        return response()->json("DELETED", 204);
+    }
+
+    public function checkIfSuperrior($type) {
+        if(auth('api')->user()->type<=$type){
+            return true;
+        }else{
+            return false;
+        }
     }
 }
